Make manage category work.
